- Setup Api service with retrofit - Added RemoteModel classes - added ApiHelper.kt class to manage data response - did pull from `development` - Moved Remote Module to remote datasource
