+++ CalculatorApp/app/src/main/java/com/thalajaat/calyxcalculator/utils/ResponseState.kt
@@ -1,10 +1,22 @@
 package com.thalajaat.calyxcalculator.utils
 
-import com.thalajaat.calyxcalculator.network.models.CurrencyConversionResponse
-import com.thalajaat.calyxcalculator.network.models.ErrorResponse
+import com.thalajaat.calyxcalculator.data.datasources.remote.models.CurrencyConversionResponse
+import com.thalajaat.calyxcalculator.data.datasources.remote.models.ErrorResponse
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.catch
+import kotlinx.coroutines.flow.onStart
 
-sealed class ResponseState {
-    data class Success(val data: CurrencyConversionResponse): ResponseState()
-    data class Error(val error : ErrorResponse): ResponseState()
-    object Loading: ResponseState()
+sealed class ResponseState<T>(val data: T? = null, val message: String? = null) {
+    class Success<T>(data: T?): ResponseState<T>(data)
+    class Error<T>(message : String?, data: T? = null): ResponseState<T>(data ,message)
+    class Loading<T>(data : T? = null) : ResponseState<T>(data)
+}
+
+fun<T> Flow<ResponseState<T>>.handleError(): Flow<ResponseState<T>>{
+    return onStart {
+        emit(ResponseState.Loading(null))
+    }.catch {
+
+    }
+
 }
